requirements.txt
numpy==1.24.2
scipy==1.11.1
scikit-image==0.21.0
Pillow==9.4.0
